requirements.txt
aldryn-apphooks-config==0.3.1
aldryn-boilerplates==0.7.5
aldryn-bootstrap3==1.2.1
aldryn-categories==1.0.4
aldryn-common==1.0.4
aldryn-newsblog==1.3.1
aldryn-people==1.2.2
aldryn-reversion==1.1.0
aldryn-search==0.3.0
aldryn-style==1.1.4
aldryn-translation-tools==0.2.1
altgraph==0.10.2
apipkg==1.4
asgiref==1.0.0
attrs==16.3.0
autobahn==0.17.2
Automat==0.5.0
Babel==2.3.4
backport-collections==0.1
bdist-mpkg==0.5.0
bonjour-py==0.3
channels==1.0.3
cmsplugin-filer==1.1.1
constantly==15.1.0
coverage==4.3.4
daphne==1.0.3
Django==1.8
django-appconf==1.0.2
django-appdata==0.1.6
django-category==1.9
django-classy-tags==0.8.0
django-cms==3.4.2
django-durationfield==0.5.3
django-filer==1.2.2
django-formtools==2.0
django-haystack==2.6.0
django-likes==1.10
django-meta==1.3.2
django-meta-mixin==0.3.0
django-mptt==0.8.7
django-object-tools==1.9
django-parler==1.7
django-phonenumber-field==1.1.0
django-photologue===2.7.praekelt
django-polymorphic==0.8.1
django-preferences==0.1
django-publisher==0.0.3
django-reversion==1.10.2
django-secretballot==0.6.0
django-sekizai==0.10.0
django-sites-groups==1.9.2
django-sortedm2m==1.3.3
django-spurl==0.6.4
django-standard-form==1.1.1
django-taggit==0.22.0
django-taggit-autosuggest==0.3.0
django-taggit-templatetags==0.2.5
django-tastypie==0.13.3
django-templatetag-sugar==1.0
django-treebeard==4.1.0
djangocms-admin-style==1.2.6.2
djangocms-apphook-setup==0.1.2
djangocms-attributes-field==0.1.2
djangocms-blog==0.8.10
djangocms-column==1.7.0
djangocms-googlemap==1.0.2
djangocms-link==2.1.0
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.4.0
djangocms-video==2.0.3
easy-thumbnails==2.3
execnet==1.4.1
Flask==0.10.1
Flask-SocketIO==2.8.4
greenlet==0.4.2
gunicorn==18.0
haikunator==2.1.0
html5lib==0.9999999
incremental==16.10.1
itsdangerous==0.23
Jinja2==2.7.2
jmbo==0.5.5
lxml==3.7.3
macholib==1.5.1
MarkupSafe==0.18
matplotlib==1.3.1
modulegraph==0.10.4
MySQL-python==1.2.5
numpy==1.8.0rc1
olefile==0.44
phonenumbers==8.3.2
phonenumberslite==8.3.2
Pillow==4.0.0
pip==9.0.1
py==1.4.32
py2app==0.7.3
pycrypto==2.6.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
pytest==3.0.6
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-xdist==1.15.0
python-dateutil==1.5
python-engineio==1.2.2
python-mimeparse==1.6.0
python-slugify==1.2.0
python-socketio==1.7.1
pytz==2013.7
runp==0.0.2
scipy==0.13.0b1
setuptools==18.5
six==1.10.0
South==1.0.2
Twisted==13.1.0
txaio==2.6.1
Unidecode==0.4.20
URLObject==2.4.2
vboxapi==1.0
virtualenv==15.1.0
Werkzeug==0.9.4
xattr==0.6.4
YURL==0.13
zope.interface==4.1.1
